requirements.txt
fastapi
uvicorn
python-dotenv
faiss-cpu
sentence-transformers
rank-bm25
transformers
torch
requests
redis
python-jose
passlib[bcrypt]
ragas
datasets
python-multipart

